Fixed bug with game won at beginning of game

+++ ProjectContext2018Unity/Assets/Scripts/Menus/Game Over/PlayerInfoItem.cs
@@ -12,7 +12,7 @@ namespace UI {
         public void Init(int index, Player player) {
             indexTxt.text = index.ToString();
             nameTxt.text = player.Name;
-            scoreTxt.text = player.ScoreManager.Score.ToString();
+            scoreTxt.text = ((int)player.ScoreManager.Score).ToString();
             moneyTxt.text = player.ScoreManager.money.ToString();
             populationTxt.text = player.ScoreManager.inhabitants.ToString();
         }
